Service time does not differentiate between activities
I changed the service time such that it does not require an activity to be passed. That limits the applicability to only extract the service time for individual activities, but just calculating the service times for each event might be much more of a regular case. An individual activity feature could also be added (although the problem of how to fill missing values for events that are not of this activity would still be the same). I removed event_duration due to redudancy